feat(pages/projects/add): Implementation of static project creation
